refactor: MenuServiceTest -> MenuRepositoryTest로 리팩터링
